refactor: post-review cleanups (error-path session/file cleanup, 204 ordering, dead field, import paths)
